+++ src/mako/http/response/Status.php
@@ -8,14 +8,18 @@
 namespace mako\http\response;
 
 use Deprecated;
+use mako\http\response\traits\StatusTrait;
+use Override;
 
 /**
  * HTTP status codes.
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status
  */
-enum Status: int
+enum Status: int implements StatusInterface
 {
+	use StatusTrait;
+
 	// 1xx Informational
 
 	case Continue = 100;
@@ -95,16 +99,9 @@ enum Status: int
 	case NetworkAuthenticationRequired = 511;
 
 	/**
-	 * Returns the status code.
+	 * {@inheritDoc}
 	 */
-	public function getCode(): int
-	{
-		return $this->value;
-	}
-
-	/**
-	 * Returns the status message.
-	 */
+	#[Override]
 	public function getMessage(): string
 	{
 		return match ($this) {
@@ -185,45 +182,5 @@ enum Status: int
 			self::NotExtended => 'Not Extended',
 			self::NetworkAuthenticationRequired => 'Network Authentication Required',
 		};
-	}
-
-	/**
-	 * Returns TRUE if the status is informational and FALSE if not.
-	 */
-	public function isInformational(): bool
-	{
-		return $this->value >= 100 && $this->value <= 199;
-	}
-
-	/**
-	 * Returns TRUE if the status is a success and FALSE if not.
-	 */
-	public function isSuccess(): bool
-	{
-		return $this->value >= 200 && $this->value <= 299;
-	}
-
-	/**
-	 * Returns TRUE if the status is a redirect and FALSE if not.
-	 */
-	public function isRedirect(): bool
-	{
-		return $this->value >= 300 && $this->value <= 399;
-	}
-
-	/**
-	 * Returns TRUE if the status is a client error and FALSE if not.
-	 */
-	public function isClientError(): bool
-	{
-		return $this->value >= 400 && $this->value <= 499;
-	}
-
-	/**
-	 * Returns TRUE if the status is a server error and FALSE if not.
-	 */
-	public function isServerError(): bool
-	{
-		return $this->value >= 500 && $this->value <= 599;
 	}
 }