+++ src/mako/http/response/senders/Redirect.php
@@ -11,10 +11,9 @@ use mako\http\exceptions\HttpException;
 use mako\http\Request;
 use mako\http\Response;
 use mako\http\response\Status;
+use mako\http\response\StatusInterface;
 use Override;
 
-use function in_array;
-use function is_int;
 use function sprintf;
 
 /**
@@ -48,27 +47,16 @@ class Redirect implements ResponseSenderInterface
 	public const Status PERMANENT_REDIRECT = Status::PermanentRedirect;
 
 	/**
-	 * Supported redirect types.
-	 */
-	public const array SUPPORTED_STATUS_CODES = [
-		Status::MovedPermanently,
-		Status::Found,
-		Status::SeeOther,
-		Status::TemporaryRedirect,
-		Status::PermanentRedirect,
-	];
-
-	/**
 	 * Status code.
 	 */
-	protected Status $status;
+	protected StatusInterface $status;
 
 	/**
 	 * Constructor.
 	 */
 	public function __construct(
 		protected string $location,
-		int|Status $status = Status::Found
+		StatusInterface $status = Status::Found
 	) {
 		$this->setStatus($status);
 	}
@@ -78,12 +66,10 @@ class Redirect implements ResponseSenderInterface
 	 *
 	 * @return $this
 	 */
-	public function setStatus(int|Status $status): Redirect
+	public function setStatus(StatusInterface $status): Redirect
 	{
-		$status = is_int($status) ? Status::from($status) : $status;
-
-		if (!in_array($status, self::SUPPORTED_STATUS_CODES)) {
-			throw new HttpException(sprintf('Unsupported redirect status code [ %s ].', $status->value));
+		if (!$status->isRedirect()) {
+			throw new HttpException(sprintf('Unsupported redirect status code [ %s ].', $status->getCode()));
 		}
 
 		$this->status = $status;
@@ -154,7 +140,7 @@ class Redirect implements ResponseSenderInterface
 	/**
 	 * Returns the HTTP status.
 	 */
-	public function getStatus(): Status
+	public function getStatus(): StatusInterface
 	{
 		return $this->status;
 	}