+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 	- `Status::isRedirect()`
 	- `Status::isClientError()`
 	- `Status::isServerError()`
+* It is now possible to set non-standard HTTP statuses using the `CustomStatus` class.
 
 #### Changes
 
@@ -48,6 +49,7 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 * Removed non-standard cases from the `Status` enum:
 	- `InvalidToken`
 	- `TokenRequired`
+* You can no longer use an integer to set the response status code. A `Status` enum case or `StatusInterface` implementation must be used instead.
 
 #### Deprecations